+++ src/components/CapacityBar.js
