+++ src/components/ChartControl.js
