+++ src/pages/PublicView.js
